(fmap snd globDir) is now nubbed.
Data.Set is used: while this adds a dependency on containers, it's the
only sane option since we want to preserve ordering. Data.List.nub is
out of the question since it's O(nm): the Set approach is O(n log m).
(Where n is the length of the list to be nubbed and m is the number of
unique symbols in it.)