./foo is equivalent to foo: handled and tested.
This needs handling in both Directory.separate and Match.begMatch.

An alternative would have been to handle it only in the optimizer and
rely on the fact that a user can never get hold of an unoptimized
pattern. I'd rather take a little performance hit than risk breaking
this later on, and the optimizer really shouldn't affect the behaviour
of things.