Handle open ranges as Strings, not Integers.
Fixes bug: <->3 did not match 123 since the <-> ate the whole number. A
few regression tests were added for this.

In addition, this should be faster than the previous way; at least it
seems to perform less allocations.

Previously <a-b> where a > b was simply turned into <b-a>: since we
don't now compare the two strings at compile time (which could be
expensive) we simply ignore this case. Thus such an <a-b> never matches
anything, but that's the user's fault.

Also, turning <a-a> into a literal is now done in the optimizer, as it
should be.