Fix `commonDirectory`, fixes #18
Make sure the correct number of directory separators remain present in
the patterns returned by `commonDirectory`.

Also fixes another issue where commonDirectory was not correctly
handling the character range [.] at the beginning of a pattern or
directly after a path separator; since dots at the beginning of
filenames are required to match explicitly we do not allow this pattern
to match anything.